requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.2.3
django-cors-headers==4.6.0
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==5.2.0
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
